VFS-5749 Refactored key-value helper batch deletion

+++ src/keyValueAdapter.cc
@@ -85,7 +85,7 @@ folly::Future<folly::IOBufQueue> KeyValueFileHandle::read(
 
     // NOLINTNEXTLINE(clang-analyzer-cplusplus.NewDelete)
     return folly::via(
-        m_executor.get(), [ this, offset, size, self = shared_from_this() ] {
+        m_executor.get(), [this, offset, size, self = shared_from_this()] {
             return readBlocks(offset, size);
         });
 }
@@ -99,7 +99,7 @@ folly::Future<std::size_t> KeyValueFileHandle::write(
         return folly::makeFuture<std::size_t>(0);
 
     return folly::via(m_executor.get(),
-        [ this, offset, buf = std::move(buf), self = shared_from_this() ] {
+        [this, offset, buf = std::move(buf), self = shared_from_this()] {
             const auto size = buf.chainLength();
             if (size == 0)
                 return folly::makeFuture<std::size_t>(0);
@@ -115,19 +115,20 @@ folly::Future<std::size_t> KeyValueFileHandle::write(
                 const auto blockSize = std::min<std::size_t>(
                     m_blockSize - blockOffset, size - bufOffset);
 
-                auto writeFuture = via(m_executor.get(), [
-                    this, iobuf = buf.front()->clone(), blockId, blockOffset,
-                    bufOffset, blockSize, self = shared_from_this()
-                ]() mutable {
-                    folly::IOBufQueue bufq{
-                        folly::IOBufQueue::cacheChainLength()};
+                auto writeFuture = via(m_executor.get(),
+                    [this, iobuf = buf.front()->clone(), blockId, blockOffset,
+                        bufOffset, blockSize,
+                        self = shared_from_this()]() mutable {
+                        folly::IOBufQueue bufq{
+                            folly::IOBufQueue::cacheChainLength()};
 
-                    bufq.append(std::move(iobuf));
-                    bufq.trimStart(bufOffset);
-                    bufq.trimEnd(bufq.chainLength() - blockSize);
+                        bufq.append(std::move(iobuf));
+                        bufq.trimStart(bufOffset);
+                        bufq.trimEnd(bufq.chainLength() - blockSize);
 
-                    return writeBlock(std::move(bufq), blockId, blockOffset);
-                });
+                        return writeBlock(
+                            std::move(bufq), blockId, blockOffset);
+                    });
 
                 writeFutures.emplace_back(std::move(writeFuture));
 
@@ -172,21 +173,45 @@ folly::Future<folly::Unit> KeyValueAdapter::unlink(
     if (currentSize == 0)
         return folly::makeFuture();
 
-    return folly::via(m_executor.get(),
-        [ fileId, helper = m_helper, blockSize = m_blockSize, currentSize ] {
-            try {
-                folly::fbvector<folly::fbstring> keysToDelete;
-                for (size_t objectId = 0; objectId <= (currentSize / blockSize);
-                     objectId++) {
-                    keysToDelete.emplace_back(helper->getKey(fileId, objectId));
-                }
+    // Calculate the list of objects into which the file has been
+    // split on the storage
+    folly::fbvector<folly::fbstring> keysToDelete;
+    for (size_t objectId = 0; objectId <= (currentSize / m_blockSize);
+         objectId++) {
+        keysToDelete.emplace_back(m_helper->getKey(fileId, objectId));
+    }
+
+    // In case the storage supports batch delete (in a single request)
+    // use it, otherwise schedule deletions of each object individually
+    // in asynchronous requests
+    if (m_helper->supportsBatchDelete()) {
+        return folly::via(m_executor.get(),
+            [keysToDelete = std::move(keysToDelete), helper = m_helper] {
                 helper->deleteObjects(keysToDelete);
-            }
-            catch (const std::system_error &e) {
-                logError("unlink", e);
-                throw;
-            }
-        });
+            });
+    }
+
+    auto batchIndex = 0ul;
+    do {
+        folly::fbvector<folly::Future<folly::Unit>> futs;
+        futs.reserve(MAX_ASYNC_DELETE_OBJECTS);
+
+        for (auto i = batchIndex;
+             i < std::min(batchIndex + MAX_ASYNC_DELETE_OBJECTS,
+                     keysToDelete.size());
+             i++) {
+            futs.emplace_back(folly::via(m_executor.get(),
+                [keyToDelete = keysToDelete.at(i), helper = m_helper] {
+                    helper->deleteObject(keyToDelete);
+                }));
+        }
+
+        folly::collectAll(futs.begin(), futs.end()).get();
+
+        batchIndex += MAX_ASYNC_DELETE_OBJECTS;
+    } while (batchIndex < MAX_ASYNC_DELETE_OBJECTS);
+
+    return folly::makeFuture();
 }
 
 folly::Future<folly::Unit> KeyValueAdapter::truncate(
@@ -197,62 +222,83 @@ folly::Future<folly::Unit> KeyValueAdapter::truncate(
     if (static_cast<size_t>(size) == currentSize)
         return folly::makeFuture();
 
-    return folly::via(m_executor.get(), [
-        fileId, size, currentSize, helper = m_helper, locks = m_locks,
-        defBlockSize = m_blockSize
-    ] {
-        try {
-            auto currentLastBlockId = getBlockId(currentSize, defBlockSize);
-            auto newLastBlockId = getBlockId(size, defBlockSize);
-            auto newLastBlockOffset = getBlockOffset(size, defBlockSize);
+    auto currentLastBlockId = getBlockId(currentSize, m_blockSize);
+    auto newLastBlockId = getBlockId(size, m_blockSize);
+    auto newLastBlockOffset = getBlockOffset(size, m_blockSize);
 
-            // Generate a list of objects to delete
-            folly::fbvector<folly::fbstring> keysToDelete;
-            for (size_t objectId = 0; objectId <= currentLastBlockId;
-                 objectId++) {
-                if ((objectId > newLastBlockId) ||
-                    ((objectId == newLastBlockId) &&
-                        (newLastBlockOffset == 0))) {
-                    keysToDelete.emplace_back(helper->getKey(fileId, objectId));
-                }
-            }
+    // Generate a list of objects to delete
+    folly::fbvector<folly::fbstring> keysToDelete;
+    for (size_t objectId = 0; objectId <= currentLastBlockId; objectId++) {
+        if ((objectId > newLastBlockId) ||
+            ((objectId == newLastBlockId) && (newLastBlockOffset == 0))) {
+            keysToDelete.emplace_back(m_helper->getKey(fileId, objectId));
+        }
+    }
 
-            auto key = helper->getKey(fileId, newLastBlockId);
-            auto remainderBlockSize =
-                static_cast<std::size_t>(newLastBlockOffset);
+    auto key = m_helper->getKey(fileId, newLastBlockId);
+    auto remainderBlockSize = static_cast<std::size_t>(newLastBlockOffset);
+    if (remainderBlockSize == 0 && newLastBlockId > 0) {
+        key = m_helper->getKey(fileId, newLastBlockId - 1);
+        remainderBlockSize = m_blockSize;
+    }
 
-            if (remainderBlockSize == 0 && newLastBlockId > 0) {
-                key = helper->getKey(fileId, newLastBlockId - 1);
-                remainderBlockSize = defBlockSize;
-            }
-
+    return folly::via(m_executor.get(),
+        [fileId, remainderBlockSize, newLastBlockId, key, helper = m_helper,
+            locks = m_locks] {
             if (remainderBlockSize > 0 || newLastBlockId > 0) {
                 Locks::accessor acc;
                 locks->insert(acc, key);
+                auto g = folly::makeGuard([&]() mutable { locks->erase(acc); });
 
                 try {
                     auto buf = fillToSize(
                         readBlock(helper, key, 0, remainderBlockSize),
                         remainderBlockSize);
                     helper->putObject(key, std::move(buf));
-                    locks->erase(acc);
                 }
                 catch (...) {
-                    locks->erase(acc);
-                    LOG(ERROR) << "Truncate failed due to unknown error during "
+                    LOG(ERROR) << "Truncate failed due to unknown "
+                                  "error during "
                                   "'fillToSize'";
                     throw;
                 }
             }
+        })
+        .then([keysToDelete = std::move(keysToDelete), helper = m_helper,
+                  executor = m_executor]() {
+            if (!keysToDelete.empty()) {
+                if (helper->supportsBatchDelete()) {
+                    helper->deleteObjects(keysToDelete);
+                }
+                else {
+                    auto batchIndex = 0ul;
+                    do {
+                        folly::fbvector<folly::Future<folly::Unit>> futs;
+                        futs.reserve(MAX_ASYNC_DELETE_OBJECTS);
 
-            if (!keysToDelete.empty())
-                helper->deleteObjects(keysToDelete);
-        }
-        catch (const std::system_error &e) {
-            logError("truncate", e);
-            throw;
-        }
-    });
+                        for (auto i = batchIndex;
+                             i < std::min(batchIndex + MAX_ASYNC_DELETE_OBJECTS,
+                                     keysToDelete.size());
+                             i++) {
+
+                            futs.emplace_back(
+                                // NOLINTNEXTLINE(clang-analyzer-cplusplus.NewDelete)
+                                folly::via(executor.get(),
+                                    [keyToDelete = keysToDelete.at(i), helper] {
+                                        helper->deleteObject(keyToDelete);
+                                        return folly::makeFuture();
+                                    }));
+                        }
+
+                        folly::collectAll(futs.begin(), futs.end()).get();
+
+                        batchIndex += MAX_ASYNC_DELETE_OBJECTS;
+                    } while (batchIndex < MAX_ASYNC_DELETE_OBJECTS);
+                }
+            }
+
+            return folly::makeFuture();
+        });
 }
 
 const Timeout &KeyValueAdapter::timeout()
@@ -284,12 +330,11 @@ folly::Future<folly::IOBufQueue> KeyValueFileHandle::readBlocks(
             std::min(static_cast<std::size_t>(m_blockSize - blockOffset),
                 static_cast<std::size_t>(size - bufOffset));
 
-        auto readFuture = via(m_executor.get(), [
-            this, blockId, blockOffset, blockSize, self = shared_from_this()
-        ] {
-            return fillToSize(
-                readBlock(blockId, blockOffset, blockSize), blockSize);
-        });
+        auto readFuture = via(m_executor.get(),
+            [this, blockId, blockOffset, blockSize, self = shared_from_this()] {
+                return fillToSize(
+                    readBlock(blockId, blockOffset, blockSize), blockSize);
+            });
 
         readFutures.emplace_back(std::move(readFuture));
         bufOffset += blockSize;
